Update config/run scripts
